+++ examples/datasets/plot_spm_faces_dataset.py
@@ -119,8 +119,8 @@ stc = apply_inverse(contrast, inverse_operator, lambda2, method,
 # plot constrast
 # Plot brain in 3D with PySurfer if available. Note that the subject name
 # is already known by the SourceEstimate stc object.
-brain = stc.plot(surface='inflated', hemi='both', subjects_dir=subjects_dir)
+brain = stc.plot(surface='inflated', hemi='both', subjects_dir=subjects_dir,
+                 transparent=True, clim='auto')
 brain.set_time(170.0)  # milliseconds
-brain.scale_data_colormap(fmin=4, fmid=6, fmax=8, transparent=True)
 brain.show_view('ventral')
 # brain.save_image('dSPM_map.png')