+++ examples/decoding/plot_decoding_spatio_temporal_source.py
@@ -147,7 +147,6 @@ stc_feat = mne.SourceEstimate(feature_weights, vertices=vertices,
                               tmin=stc.tmin, tstep=stc.tstep,
                               subject='sample')
 
-brain = stc_feat.plot(subject=subject, fmin=1, fmid=5.5, fmax=20)
+brain = stc_feat.plot(subject=subject, clim='auto')
 brain.set_time(100)
-brain.show_view('l')
-# take the medial view to further explore visual areas
+brain.show_view('l')  # take the medial view to further explore visual areas