+++ examples/inverse/plot_compute_mne_inverse.py
@@ -47,8 +47,8 @@ plt.show()
 
 # Plot brain in 3D with PySurfer if available. Note that the subject name
 # is already known by the SourceEstimate stc object.
-brain = stc.plot(surface='inflated', hemi='rh', subjects_dir=subjects_dir)
-brain.scale_data_colormap(fmin=8, fmid=12, fmax=15, transparent=True)
+brain = stc.plot(surface='inflated', hemi='rh', subjects_dir=subjects_dir,
+                 transparent=True, clim='auto')
 brain.show_view('lateral')
 
 # use peak getter to move vizualization to the time point of the peak