+++ examples/inverse/plot_mne_point_spread_function.py
@@ -64,23 +64,17 @@ stc_psf_meg, _ = point_spread_function(
 # stc_psf_eegmeg.save('psf_eegmeg')
 # stc_psf_meg.save('psf_meg')
 
-fmin = 0.
+clim = dict(kind='percent', lims=[93, 99, 100])
 time_label = "EEGMEG %d"
-fmax = stc_psf_eegmeg.data[:, 0].max()
-fmid = fmax / 2.
 brain_eegmeg = stc_psf_eegmeg.plot(surface='inflated', hemi='rh',
                                    subjects_dir=subjects_dir,
-                                   time_label=time_label, fmin=fmin,
-                                   fmid=fmid, fmax=fmax,
+                                   time_label=time_label, clim=clim,
                                    figure=mlab.figure(size=(500, 500)))
 
 time_label = "MEG %d"
-fmax = stc_psf_meg.data[:, 0].max()
-fmid = fmax / 2.
 brain_meg = stc_psf_meg.plot(surface='inflated', hemi='rh',
                              subjects_dir=subjects_dir,
-                             time_label=time_label, fmin=fmin,
-                             fmid=fmid, fmax=fmax,
+                             time_label=time_label, clim=clim,
                              figure=mlab.figure(size=(500, 500)))
 
 # The PSF is centred around the right auditory cortex label,