+++ examples/plot_ssp_projs_sensitivity_map.py
@@ -36,6 +36,6 @@ ssp_ecg_map = sensitivity_map(fwd, ch_type='grad', projs=projs, mode='angle')
 plt.hist(ssp_ecg_map.data.ravel())
 plt.show()
 
-args = dict(fmin=0.2, fmid=0.6, fmax=1., smoothing_steps=7, hemi='rh',
-            subjects_dir=subjects_dir)
+args = dict(clim=dict(kind='value', lims=(0.2, 0.6, 1.)), smoothing_steps=7,
+            hemi='rh', subjects_dir=subjects_dir)
 ssp_ecg_map.plot(subject='sample', time_label='ECG SSP sensitivity', **args)