+++ mne/viz/tests/test_3d.py
@@ -3,6 +3,7 @@
 #          Martin Luessi <mluessi@nmr.mgh.harvard.edu>
 #          Eric Larson <larson.eric.d@gmail.com>
 #          Mainak Jas <mainak@neuro.hut.fi>
+#          Mark Wronkiewicz <wronk.mark@gmail.com>
 #
 # License: Simplified BSD
 
@@ -15,7 +16,7 @@ from numpy.testing import assert_raises
 from mne import SourceEstimate
 from mne import make_field_map, pick_channels_evoked, read_evokeds
 from mne.viz import (plot_sparse_source_estimates, plot_source_estimates,
-                     plot_trans, mne_analyze_colormap)
+                     plot_trans)
 from mne.utils import requires_mayavi, requires_pysurfer
 from mne.datasets import testing
 from mne.source_space import read_source_spaces
@@ -50,17 +51,18 @@ def test_plot_sparse_source_estimates():
     n_time = 5
     n_verts = sum(len(v) for v in vertices)
     stc_data = np.zeros((n_verts * n_time))
-    stc_data[(np.random.rand(20) * n_verts * n_time).astype(int)] = 1
+    stc_size = stc_data.size
+    stc_data[(np.random.rand(stc_size / 20) * stc_size).astype(int)] = \
+        np.random.rand(stc_data.size / 20)
     stc_data.shape = (n_verts, n_time)
     stc = SourceEstimate(stc_data, vertices, 1, 1)
-    colormap = mne_analyze_colormap(format='matplotlib')
-    # don't really need to test matplotlib method since it's not used now...
-    colormap = mne_analyze_colormap()
+    colormap = 'mne_analyze'
     plot_source_estimates(stc, 'sample', colormap=colormap,
                           config_opts={'background': (1, 1, 0)},
-                          subjects_dir=subjects_dir, colorbar=True)
+                          subjects_dir=subjects_dir, colorbar=True,
+                          clim='auto')
     assert_raises(TypeError, plot_source_estimates, stc, 'sample',
-                  figure='foo', hemi='both')
+                  figure='foo', hemi='both', clim='auto')
 
     # now do sparse version
     vertices = sample_src[0]['vertno']
@@ -101,3 +103,49 @@ def test_plot_trans():
     assert_raises(ValueError, plot_trans, evoked.info, trans_fname,
                   subject='sample', subjects_dir=subjects_dir,
                   ch_type='bad-chtype')
+
+
+@testing.requires_testing_data
+@requires_pysurfer
+@requires_mayavi
+def test_limits_to_control_points():
+    """Test functionality for determing control points
+    """
+    sample_src = read_source_spaces(src_fname)
+
+    vertices = [s['vertno'] for s in sample_src]
+    n_time = 5
+    n_verts = sum(len(v) for v in vertices)
+    stc_data = np.zeros((n_verts * n_time))
+    stc_data[(np.random.rand(20) * n_verts * n_time).astype(int)] = 1
+    stc_data.shape = (n_verts, n_time)
+    stc = SourceEstimate(stc_data, vertices, 1, 1)
+
+    # Test both types of incorrect limits key (lims/pos_lims)
+    clim = dict(kind='value', lims=(5, 10, 15))
+    colormap = 'mne_analyze'
+    assert_raises(KeyError, plot_source_estimates, stc, 'sample',
+                  colormap=colormap, clim=clim)
+
+    clim = dict(kind='value', pos_lims=(5, 10, 15))
+    colormap = 'hot'
+    assert_raises(KeyError, plot_source_estimates, stc, 'sample',
+                  colormap=colormap, clim=clim)
+
+    # Test for correct clim values
+    clim['pos_lims'] = (5, 10, 15, 20)
+    colormap = 'mne_analyze'
+    assert_raises(ValueError, plot_source_estimates, stc, 'sample',
+                  colormap=colormap, clim=clim)
+    clim = 'foo'
+    assert_raises(ValueError, plot_source_estimates, stc, 'sample',
+                  colormap=colormap, clim=clim)
+    clim = (5, 10, 15)
+    assert_raises(ValueError, plot_source_estimates, stc, 'sample',
+                  colormap=colormap, clim=clim)
+
+    # Test that stc.data contains enough unique values to use percentages
+    clim = 'auto'
+    stc._data = np.zeros_like(stc.data)
+    assert_raises(ValueError, plot_source_estimates, stc, 'sample',
+                  colormap=colormap, clim=clim)